pythonBackend/requirements.txt
Flask
pandas
requests
beautifulsoup4
numpy
matplotlib
flask-cors
scikit-learn
textblob 
nltk 
newspaper3k 
transformers 
sentencepiece 
python-dateutil
vaderSentiment